final fix

+++ bank.py
@@ -58,13 +58,13 @@ class Bank:
     def set_cvv(self, cvv):
         self.__card_cvv = cvv
 
-    def get_balance_byn(self) -> int:
+    def get_balance_byn(self) -> float:
         return self.__card_balance_byn
 
     def set_balance_byn(self, balance: int):
         self.__card_balance_byn = balance
 
-    def get_balance_usd(self) -> int:
+    def get_balance_usd(self) -> float:
         return self.__card_balance_usd
 
     def set_balance_usd(self, balance: int):
